Added a Nat-NG pattern component (http://drupal.org/project/nat_ng).
Example usage:

  <nat_ng>
	<type>Image</type>
	<vocabs>
		<vocab>Wallpaper</vocab>
		<vocab>Photo</vocab>
	</vocabs>
	<associate-node-body-with-term-description>1</associate-node-body-with-term-description>
	<delete-node-if-term-deleted>1</delete-node-if-term-deleted>
	<delete-natng-node-if-natng-term-deleted>1</delete-natng-node-if-natng-term-deleted>
	<allow-users-to-define-synonyms>0</allow-users-to-define-synonyms>
	<natng-term-in-node-view-points-to-node>1</natng-term-in-node-view-points-to-node>
  </nat_ng>

The above configuration associates the Image node type with the Wallpaper and Photo vocabularies. Thus, every time an 'Image' is created a term will be added to the 'Wallpaper' and 'Photo' vocabularies respectively.
The other tags specify additional Nat-NG settings for this node type. They correlate with the options given on the Nat-NG settings page for each node type (/admin/settings/nat-ng/settings).