Added a 'translate' attribute for  the Menu Trails pattern. If you have deployed multiple sites in different languages but with each having the same basic menu structure (albeit translated menu items) you don't want to have to repeat Menu Trail patterns for each language.
Example:

    <menu_trail>
      <menu>
        <parent>Primary links</parent>
        <item translate="true">Blogs</item>
      </menu>
      <type>Blogs</type>
    </menu_trail>

Internally when looking up the correct menu item to assign nodes of type Blogs to, the t() method is called with the 'Blogs' as the parameter and the language code set to the active language of the site. If the 'translate' attribute is ommitted then no translation is performed.